utilities

+++ contracts/interfaces/ICERC20.sol
@@ -8,15 +8,19 @@ interface ICERC20 {
 
     function supplyRatePerBlock() external returns (uint256);
 
-    function redeem(uint) external returns (uint);
+    function redeem(uint256) external returns (uint256);
 
-    function redeemUnderlying(uint) external returns (uint);
-    
+    function redeemUnderlying(uint256) external returns (uint256);
+
     function balanceOf(address account) external view returns (uint256);
 
-    function transfer(address dst, uint amount) external returns (bool);
+    function transfer(address dst, uint256 amount) external returns (bool);
 
-    function transferFrom(address src, address dst, uint amount) external returns (bool);
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 amount
+    ) external returns (bool);
 
-    function approve(address spender, uint amount) external returns (bool);
+    function approve(address spender, uint256 amount) external returns (bool);
 }
